db: don't attempt to delete model with no PK
Doing so causes an exception. Rather, if the model does not have a PK
(is not saved in the DB) do nothing.

+++ src/gentoo_build_publisher/db.py
@@ -144,7 +144,8 @@ class BuildDB:
 
     def delete(self):
         """Delete this Build from the db"""
-        self.model.delete()
+        if self.model.pk is not None:
+            self.model.delete()
 
     @classmethod
     def create(
